Make page responsive
