+++ frontend/src/browser/Browser.css
@@ -3,16 +3,6 @@
 
     padding-inline: 2rem;
     box-sizing: border-box;
-
-    display: grid;
-    grid-template-columns: 20vw 1fr;
-    grid-template-rows: 1fr;
-    grid-template-areas:    "filter result";
-
-    column-gap: 2rem;
-
-    justify-content: center;
-    align-content: center;
 }
 
 .browser-result {
@@ -29,6 +19,16 @@
     grid-area: filter;
 }
 
-.browser-filter-element {
-    margin-bottom: 2rem;
+@media screen and (min-width: 1200px) {
+    .browser {
+        display: grid;
+        grid-template-columns: 20vw 1fr;
+        grid-template-rows: 1fr;
+        grid-template-areas:    "filter result";
+
+        column-gap: 2rem;
+
+        justify-content: center;
+        align-content: center;
+    }
 }
