+++ frontend/src/browser/Browser.tsx
@@ -3,8 +3,7 @@ import { DrinkGrid } from '../drinks/DrinkGrid';
 import { Query, useSearch } from './search';
 
 import { IIngredient } from '../../../types';
-import { FilterElement, ValueSlider } from './BrowserFilter';
-import { IngredientList } from '../searchform/IngredientList';
+import { BrowserFilterPane } from './BrowserFilter';
 
 import './Browser.css';
 
@@ -23,12 +22,12 @@ export function Browser(): JSX.Element {
     }, [maxAlcoholConcentration, ingredients]);
 
     return <div className="browser">
-        <BrowserResult />
         <BrowserFilterPane 
             maxAlcoholConcentration={maxAlcoholConcentration} 
             onMaxAlcoholConcentrationUpdate={setMaxAlcoholConcentration}
             ingredients={ingredients} 
             onIngredientsUpdate={setIngredients} />
+        <BrowserResult />
     </div>;
 }
 
@@ -48,26 +47,4 @@ export function BrowserResult(): JSX.Element {
 
 function BrowserStatusMessage({ children }: { children: string }): JSX.Element {
     return <div className="browser-result browser-message">{children}</div>;
-}
-
-export function BrowserFilterPane({
-    maxAlcoholConcentration,
-    onMaxAlcoholConcentrationUpdate,
-    ingredients,
-    onIngredientsUpdate
-}: {
-    maxAlcoholConcentration: number,
-    onMaxAlcoholConcentrationUpdate: (value: number) => void,
-    ingredients: Array<IIngredient>,
-    onIngredientsUpdate: (value: Array<IIngredient>) => void
-}): JSX.Element {
-    return <div className="browser-filter">
-        <FilterElement label="Concentration of alcohol">
-            <ValueSlider min={0} max={1} precision={.05} 
-                value={maxAlcoholConcentration} onValueUpdate={onMaxAlcoholConcentrationUpdate} />
-        </FilterElement>
-        <FilterElement label="Ingredients">
-            <IngredientList ingredients={ingredients} setIngredients={onIngredientsUpdate} />
-        </FilterElement>
-    </div>;
 }
