+++ frontend/src/index.css
@@ -35,16 +35,23 @@ body {
 }
 
 header {
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
+  margin-bottom: 1rem;
+}
+
+header::after {
+    content: " ";
+    display: block;
+
+    margin-top: 1rem;
+
+    border: solid 1px var(--primary);
 }
 
 .app-title {
   color: var(--primary);
   font-family: var(--primary-font);
   text-align: center;
-  padding: 10px 20px;
+  padding: 10px 0;
   font-size: 35px;
   text-shadow: 2px 2px 10px var(--secondary-shadow);
 }
@@ -53,4 +60,35 @@ header {
   font-family: var(--accent-font);
   font-size: 30px;
   text-shadow: 2px 2px 10px var(--secondary-shadow);
+}
+
+.app-title::after {
+    content: " ";
+    display: block;
+
+    margin-top: .9rem;
+    margin-bottom: .3rem;
+    border: solid 1px var(--primary);
+}
+
+@media screen and (min-width: 1200px) {
+    header {
+        display: flex;
+        justify-content: space-between;
+        align-items: center;
+
+        margin-bottom: 0;
+    }
+
+    header::after {
+        all: unset;
+    }
+
+    .app-title {
+        padding-right: 20px;
+    }
+
+    .app-title::after {
+        all: unset;
+    }
 }
