+++ frontend/src/nav/Navbar.css
@@ -2,7 +2,9 @@ nav {
     flex-grow: 1;
 
     display: flex;
+    align-items: center;
     justify-content: space-around;
+    flex-wrap: wrap;
 }
 
 .nav-item {
@@ -11,6 +13,12 @@ nav {
     font-size: 1.1rem;
     cursor: pointer;
     user-select: none;
+
+    margin: .3rem;
+
+    text-align: center;
+
+    min-width: 200px;
 }
 
 .nav-item:hover {
@@ -18,18 +26,20 @@ nav {
     text-underline-offset: 0.4em;
 }
 
-nav::before, nav::after {
-    display: block;
-    content: " ";
-
-    min-height: 100%;
-
-    border: solid 1px var(--primary);
-}
-
 .nav-item.selected {
     font-weight: bold;
     text-decoration: underline;
     text-decoration-thickness: 0.1em;
     text-underline-offset: 0.3em;
+}
+
+@media screen and (min-width: 1200px) {
+    nav::before, nav::after {
+        display: block;
+        content: " ";
+    
+        border: solid 1px var(--primary);
+
+        align-self: stretch;
+    }
 }
