+++ frontend/src/tipsiness/searchform/searchform.css
@@ -1,4 +1,4 @@
-.segment {
+.searchform-segment {
     padding: 1em;
     margin: 1em 1em 1em 0em;
     background-color: var(--secondary);
@@ -6,11 +6,42 @@
     box-shadow: 2px 2px 10px var(--secondary-shadow);
 }
 
+.searchform {
+    max-width: 1000px;
+    margin: 0 auto;
+}
+
+@media screen and (min-width: 1200px) {
+    .searchform {
+        display: grid;
+        grid-template-columns: max-content 2fr;
+        grid-template-areas:    "selector weight"
+                                "selector fridge"
+                                "submit submit";
+    }
+}
+
+.searchform-segment-selector {
+    grid-area: selector;
+}
+
+.searchform-segment-weight {
+    grid-area: weight;
+}
+
+.searchform-segment-fridge {
+    grid-area: fridge;
+}
+
+.searchform > button {
+    grid-area: submit;
+    max-width: 300px;
+}
+
 .searchform-label {
     font-size: 20px;
     font-weight: bold;
 }
-
 
 .searchform-input-container {
     display: flex;
@@ -20,7 +51,6 @@
 .searchform-input {
     all: unset;
     background-color: var(--primary-opacity);
-    width: 90%;
     flex-grow: 1;
     font-size: 20px;
     padding: 5px 10px 5px 10px;
